Evaluate 복합키 제거, 별도의 컬럼으로 관리
