+++ src/main/java/com/depromeet/health/exception/AlreadyEvaluatedUserException.java
@@ -1,0 +1,4 @@
+package com.depromeet.health.exception;
+
+public class AlreadyEvaluatedUserException extends Exception {
+}
