+++ src/main/java/com/depromeet/health/model/Evaluate.java
@@ -1,28 +1,50 @@
 package com.depromeet.health.model;
 
-import java.io.Serializable;
-
-import javax.persistence.EmbeddedId;
 import javax.persistence.Entity;
+import javax.persistence.GeneratedValue;
+import javax.persistence.GenerationType;
+import javax.persistence.Id;
 
 @Entity
-public class Evaluate implements Serializable {
-    @EmbeddedId
-    EvaluateId id;
+public class Evaluate {
+    @Id
+    @GeneratedValue(strategy = GenerationType.IDENTITY)
+    private Long id;
+
+    private Long userId;
+
+    private Long postId;
 
     public Evaluate() {
-
     }
 
-    public Evaluate(EvaluateId id) {
-        this.id = id;
+
+    public Evaluate(Long userId, Long postId) {
+        this.userId = userId;
+        this.postId = postId;
     }
 
-    public EvaluateId getId() {
+    public Long getId() {
         return id;
     }
 
-    public void setId(EvaluateId id) {
+    public void setId(Long id) {
         this.id = id;
+    }
+
+    public Long getUserId() {
+        return userId;
+    }
+
+    public void setUserId(Long userId) {
+        this.userId = userId;
+    }
+
+    public Long getPostId() {
+        return postId;
+    }
+
+    public void setPostId(Long postId) {
+        this.postId = postId;
     }
 }
