--- src/main/java/com/depromeet/health/model/EvaluateId.java
@@ -1,51 +1,0 @@
-package com.depromeet.health.model;
-
-import java.io.Serializable;
-import java.util.Objects;
-
-import javax.persistence.Embeddable;
-
-@Embeddable
-public class EvaluateId implements Serializable {
-    private Long userId;
-
-    private Long postId;
-
-    public EvaluateId() {
-    }
-
-    public EvaluateId(Long userId, Long postId) {
-        this.userId = userId;
-        this.postId = postId;
-    }
-
-    public Long getUserId() {
-        return userId;
-    }
-
-    public void setUserId(Long userId) {
-        this.userId = userId;
-    }
-
-    public Long getPostId() {
-        return postId;
-    }
-
-    public void setPostId(Long postId) {
-        this.postId = postId;
-    }
-
-    @Override
-    public boolean equals(Object o) {
-        if (this == o) return true;
-        if (o == null || getClass() != o.getClass()) return false;
-        EvaluateId that = (EvaluateId) o;
-        return Objects.equals(userId, that.userId) &&
-                Objects.equals(postId, that.postId);
-    }
-
-    @Override
-    public int hashCode() {
-        return Objects.hash(userId, postId);
-    }
-}
