+++ src/main/java/com/depromeet/health/repository/EvaluateRepository.java
@@ -1,8 +1,9 @@
 package com.depromeet.health.repository;
 
 import com.depromeet.health.model.Evaluate;
-import com.depromeet.health.model.EvaluateId;
+import java.util.Optional;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface EvaluateRepository extends JpaRepository<Evaluate, EvaluateId> {
+public interface EvaluateRepository extends JpaRepository<Evaluate, Long> {
+    Optional<Evaluate> findByUserIdAndPostId(Long userId, Long postId);
 }
