+++ src/main/java/com/depromeet/health/service/EvaluateService.java
@@ -1,7 +1,6 @@
 package com.depromeet.health.service;
 
 import com.depromeet.health.model.Evaluate;
-import com.depromeet.health.model.EvaluateId;
 import com.depromeet.health.repository.EvaluateRepository;
 import java.util.Optional;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -14,12 +13,12 @@ public class EvaluateService {
     EvaluateRepository evaluateRepository;
 
     public Evaluate createEvaluate(Long userId, Long postId) {
-        Evaluate evaluate = new Evaluate(new EvaluateId(userId, postId));
+        Evaluate evaluate = new Evaluate(userId, postId);
         return evaluateRepository.save(evaluate);
     }
-    
+
     public Boolean isEvaluatePost(Long userId, Long postId) {
-        Optional<Evaluate> evaluate = evaluateRepository.findById(new EvaluateId(userId, postId));
+        Optional<Evaluate> evaluate = evaluateRepository.findByUserIdAndPostId(userId, postId);
         return evaluate.isPresent();
     }
 }
